requirements.txt
appdirs==1.4.4
backcall==0.2.0
Box2D==2.3.10
box2d-py==2.3.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cycler==0.11.0
Cython==0.29.34
debugpy==1.6.7
decorator==5.1.1
docker-pycreds==0.4.0
entrypoints==0.4
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
gym==0.21.0
highway-env==1.5
idna==3.4
importlib-metadata==4.13.0
ipykernel==6.16.2
ipython==7.34.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.4
markdown-it-py==2.2.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.1
pandas==1.3.5
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickle5==0.0.12
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==4.22.4
psutil==5.9.5
ptyprocess==0.7.0
pygame==2.4.0
pyglet==1.5.27
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.30.0
rich==13.3.5
-e scikit-learn
scipy==1.7.3
seaborn==0.12.2
sentry-sdk==1.22.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
-e stable-baselines3
threadpoolctl==3.1.0
torch==1.13.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.15.2
wcwidth==0.2.6
zipp==3.15.0
